Adding support for RLD.
A good chunk of the rear font is there
Ability to switch between left and right italics (code change only for now)
Will add left and right leaning rear fonts, to match the lean.

Currently only scrolling across half of the rear display.  Will be fixed in a later update