More Serial UART comm changes
The latest changes should resolve the serial issues when talking to
the board via UART.

The issue was that with both USB and UART serial active without a
compile switch, it's possible for both to try to grab bytes and
interfere with each other.  This change locks out the USB routine
when UART data is detected which prevents the potential crosstalk.

This should get the serial working reliably now.