Fix the Teensy Serial issues.
The Teensy uses Serial3 (not Serial1) so update to use the correct HW port.
Also call the serialEvent() directly from the main loop().  While this should
not be required, as the code should be called whenever the serial interrupt is
called, the Teensy is not triggering this code.  I will keep checking and
hopefully get a better solution here, but for now it allows the code to work!