Updated the Scrolling Message Behaviour.  Other cleanup
The Scrolling message code is updated to allow a number of times the
message will be looped, rather than just looping once.

This allows the MarcDuino R2Touch behaviour to work as expected.
The 0T100 will scroll the message 1000 times (effectively forever)
So that as long as the Logic screen is highlighted to send the message
it will keep scrolling.