Reworh Both Teensy and Zero to get a Serial Port back (or make one).
On the Teensy, the Pins used for the Adjustment jumper are actually a
HW serial port.  We reclaim this to get Serial1 back on the Teensy.

On the Zero, the actual HW Serial's aren't pulled out.  That's ok, we can just
re-assign the pins, and use all the HW clockign to manage them.  Theres limited
pins we can use but this should work.

The goal is to allow the new PSIPro to be controlled from the RSeries just like
the old Teeces PSI's were.  This commit is step one in that process.

Note it's all hooked up but not yet tested.