Updates to the Zero Sketch for PSI-Pro Support
the latest update is now tested to verify correct Output of the
PSI Pro command forwarding using Pin 10 as the TX pin.

This latest update is now verified working with both Teensy and Zero
versions of the R-Series to allow the PSI-Pro to be commanded by the
R-Series.

Please note that a Logic Level Shifter may be required based on the board
you are using!