Echo the incoming command to the Tx on the Serial Port.
This allows you to use the existing Tx/Rx Serial connection to receive
a command on the Logic, then echo that command on the same port.

The intent would be to have a different device on the end of the Tx
such that you command a new device.  This in addition to the recent
changes to allow a wholly new Serial port on Pin 10/12 for the
Zero gives effectively double the Serial ports!